requirements.txt
streamlit
numpy
pandas
langchain
python-dotenv
scikit-learn
openai
tabulate